pyproject.toml
[tool.poetry]
name = "onnigiri"
version = "0.1.0"
description = "onnx-divider"
authors = ["Idein Inc."]
license = "Apache-2.0"
readme = "README.md"
keywords = ["onnx"]

[tool.poetry.dependencies]
python = "^3.8"
onnx = "^1.9.0"
onnx-simplifier = "^0.3.6"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.9.1", extras = ["lint"]}

[tool.poetry.scripts]
onnigiri = "onnigiri.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 128
py_version = "py38"

[[tool.pysen.lint.mypy_targets]]
  paths = ["."]
